requirements.txt
imaplib2==3.6
python-dotenv==1.0.1

